requirements.txt
pillow>=9.0.0
numpy>=1.20.0
tqdm>=4.60.0
pymongo>=4.0.0
requests>=2.25.0
